Updated Gui
Have Login working. So that when the login button is pressed with correct Username, Password, and Role. Once logged in, a new window is opened StartGui=Start, but StartGui has a modified function called loginGui, so that gui can work with login.  Next we will have to find a way to map username and password to a students name.